Fixed eraser tool having inconsistencies when erasing parts of a Line. Refactored function x,y and r to be int instead of unnecessarily being double's

+++ src/client/java/org/untitled/immolation/client/gui/PersistentLines.java
@@ -1,10 +1,10 @@
 package org.untitled.immolation.client.gui;
 
 public class PersistentLines {
-    public float x1, y1, x2, y2, width;
+    public int x1, y1, x2, y2, width;
     public int color;
 
-    public PersistentLines(float x1, float y1, float x2, float y2, float width, int color) {
+    public PersistentLines(int x1, int y1, int x2, int y2, int width, int color) {
         this.x1 = x1;
         this.y1 = y1;
         this.x2 = x2;
